model: queryNews function for interacting through the web server
